move private functions to the one place they are used
